Create a new maven module for the RIA JavaScript variant of the project
There's now a super project with artifact id `project4`, which contains two modules,`pure` and `ria`. Each module builds artifacts with `war` packaging that can then be used with Tomcat server. It's required to sync or reimport the maven project as the old artifacts (and modules) won't be valid anymore, due to the structural changes this commit does.

This commit also stands as an initial commit for the `ria` project variant.